fix: PGO computes optimal value for each sparsity
Previously, PGO stopped early based on whether it didn't improve on the
following iteration. However, this doesn't make sense for a nonconvex
objective function. The new behaviour replicates Halpern in the sense
that we now find the optimal vector for each `k` and then use other
techniques to pare this down